Move task to separate file
